+++ src/common/gitProvider/azureDevops.ts
@@ -3,7 +3,12 @@ import * as azdev from "azure-devops-node-api";
 import * as c from "chalk";
 import { uxLog } from "../utils";
 import { PullRequestMessageRequest, PullRequestMessageResult } from ".";
-import { CommentThreadStatus, GitPullRequestCommentThread, PullRequestAsyncStatus, PullRequestStatus } from "azure-devops-node-api/interfaces/GitInterfaces";
+import {
+  CommentThreadStatus,
+  GitPullRequestCommentThread,
+  PullRequestAsyncStatus,
+  PullRequestStatus,
+} from "azure-devops-node-api/interfaces/GitInterfaces";
 
 export class AzureDevopsProvider extends GitProviderRoot {
   private azureApi: InstanceType<typeof azdev.WebApi>;
@@ -27,11 +32,15 @@ export class AzureDevopsProvider extends GitProviderRoot {
     // Get Azure Git API
     const azureGitApi = await this.azureApi.getGitApi();
     const repositoryId = process.env.BUILD_REPOSITORY_ID || null;
+    if (repositoryId == null) {
+      uxLog(this, c.yellow("BUILD_REPOSITORY_ID must be defined"));
+      return null;
+    }
     const latestPullRequestsOnBranch = await azureGitApi.getPullRequests(repositoryId, {
       targetRefName: `refs/heads/${gitBranch}`,
-      status: PullRequestStatus.Completed
+      status: PullRequestStatus.Completed,
     });
-    const latestMergedPullRequestOnBranch = latestPullRequestsOnBranch.filter(pr => pr.mergeStatus === PullRequestAsyncStatus.Succeeded);
+    const latestMergedPullRequestOnBranch = latestPullRequestsOnBranch.filter((pr) => pr.mergeStatus === PullRequestAsyncStatus.Succeeded);
     if (latestMergedPullRequestOnBranch.length > 0) {
       const latestPullRequest = latestMergedPullRequestOnBranch[0];
       const latestPullRequestId = latestPullRequest.pullRequestId;
@@ -51,7 +60,7 @@ export class AzureDevopsProvider extends GitProviderRoot {
           }
         }
         if (deploymentCheckId) {
-          break ;
+          break;
         }
       }
     }
@@ -67,6 +76,18 @@ export class AzureDevopsProvider extends GitProviderRoot {
     const pullRequestIdStr = process.env.SYSTEM_PULLREQUEST_PULLREQUESTID || null;
     if (repositoryId == null || pullRequestIdStr == null) {
       uxLog(this, c.grey("[Azure integration] No project and pull request, so no note thread..."));
+      uxLog(
+        this,
+        c.yellow(`Following variables must be defined when available:
+- BUILD_REPOSITORY_ID
+- BUILD_BUILD_ID
+- SYSTEM_JOB_ID
+- SYSTEM_PULLREQUEST_PULLREQUESTID
+- SYSTEM_JOB_DISPLAY_NAME
+- SYSTEM_COLLECTIONURI
+- SYSTEM_TEAMPROJECT
+      `),
+      );
       return { posted: false, providerResult: { info: "No related pull request" } };
     }
     const pullRequestId = Number(pullRequestIdStr);
@@ -147,7 +168,7 @@ _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${azureJo
     return prMessage.status === "valid"
       ? CommentThreadStatus.Fixed
       : prMessage.status === "invalid"
-        ? CommentThreadStatus.Active
-        : CommentThreadStatus.Unknown;
+      ? CommentThreadStatus.Active
+      : CommentThreadStatus.Unknown;
   }
 }