+++ src/common/utils/deployUtils.ts
@@ -203,10 +203,11 @@ export async function forceSourceDeploy(
         c.cyan(`${check ? "Simulating deployment of" : "Deploying"} ${c.bold(deployment.label)} package: ${deployment.packageXmlFile} ...`),
       );
       // Try QuickDeploy
-      if (check === false && (process.env?.SFDX_HARDIS_QUICK_DEPLOY || '') !== "false") {
+      if (check === false && (process.env?.SFDX_HARDIS_QUICK_DEPLOY || "") !== "false") {
         const deploymentCheckId = await GitProvider.getDeploymentCheckId();
         if (deploymentCheckId) {
-          const quickDeployCommand = `sfdx force:source:deploy` +
+          const quickDeployCommand =
+            `sfdx force:source:deploy` +
             ` --validateddeployrequestid ${deploymentCheckId} ` +
             (options.targetUsername ? ` --targetusername ${options.targetUsername}` : "") +
             ` --wait ${process.env.SFDX_DEPLOY_WAIT_MINUTES || "60"}` +
@@ -215,14 +216,18 @@ export async function forceSourceDeploy(
           const quickDeployRes = await execSfdxJson(quickDeployCommand, commandThis, {
             output: true,
             debug: debugMode,
-            fail: false
+            fail: false,
           });
           if (quickDeployRes.status === 0) {
             uxLog(commandThis, c.green(`Successfully processed QuickDeploy for deploymentId ${deploymentCheckId}`));
             break;
-          }
-          else {
-            uxLog(commandThis, c.yellow(`Unable to perform QuickDeploy for deploymentId ${deploymentCheckId}.\n${quickDeployRes.errorMessage}\n Switching back to effective deployment.`));
+          } else {
+            uxLog(
+              commandThis,
+              c.yellow(
+                `Unable to perform QuickDeploy for deploymentId ${deploymentCheckId}.\n${quickDeployRes.errorMessage}\n Switching back to effective deployment.`,
+              ),
+            );
           }
         }
       }
@@ -733,10 +738,10 @@ export async function buildOrgManifest(targetOrgUsernameAlias, packageXmlOutputF
     // Use sfdx manifest build in current project
     await execCommand(
       `sfdx force:source:manifest:create` +
-      ` --manifestname ${manifestName}` +
-      ` --outputdir ${path.resolve(manifestDir)}` +
-      ` --includepackages managed,unlocked` +
-      ` --fromorg ${targetOrgUsernameAlias}`,
+        ` --manifestname ${manifestName}` +
+        ` --outputdir ${path.resolve(manifestDir)}` +
+        ` --includepackages managed,unlocked` +
+        ` --fromorg ${targetOrgUsernameAlias}`,
       this,
       {
         fail: true,
@@ -750,10 +755,10 @@ export async function buildOrgManifest(targetOrgUsernameAlias, packageXmlOutputF
     // Use sfdx manifest build in dummy project
     await execCommand(
       `sfdx force:source:manifest:create` +
-      ` --manifestname ${manifestName}` +
-      ` --outputdir ${path.resolve(manifestDir)}` +
-      ` --includepackages managed,unlocked` +
-      ` --fromorg ${targetOrgUsernameAlias}`,
+        ` --manifestname ${manifestName}` +
+        ` --outputdir ${path.resolve(manifestDir)}` +
+        ` --includepackages managed,unlocked` +
+        ` --fromorg ${targetOrgUsernameAlias}`,
       this,
       {
         fail: true,